added exit code management for lexing and parsing

+++ sources/builtin/builtin_unset.c
@@ -6,7 +6,7 @@
 /*   By: cscache <cscache@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/09 16:48:37 by barmarti          #+#    #+#             */
-/*   Updated: 2025/08/11 17:03:07 by cscache          ###   ########.fr       */
+/*   Updated: 2025/08/14 13:38:37 by cscache          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,8 @@
 
 static void	delete_node_env(t_env *to_delete)
 {
+	if (!to_delete)
+		return ;
 	if (to_delete->next != NULL)
 		to_delete->next->prev = to_delete->prev;
 	to_delete->prev->next = to_delete->next;
@@ -22,18 +24,21 @@ static void	delete_node_env(t_env *to_delete)
 	free(to_delete);
 }
 
-void	builtin_unset(t_env **env, char *to_delete)
+int	builtin_unset(t_env **env, char *to_delete)
 {
 	t_env	*current;
 
+	if (!env || !*env || !to_delete)
+		return (EXIT_FAILURE);
 	current = *env;
 	while (current)
 	{
 		if (!ft_strcmp(to_delete, current->key))
 		{
 			delete_node_env(current);
-			return ;
+			return (EXIT_SUCCESS);
 		}
 		current = current->next;
 	}
+	return (EXIT_SUCCESS);
 }
